moved functionality to shared.js
